DocumentID can be an Int not Long, we don't need 64 bits here and we store a lot of them

+++ worker/src/main/scala/ClusterTypes.scala
@@ -15,7 +15,7 @@ import scala.collection.mutable.ArrayBuffer
 
 object ClusterTypes {
 
-  type DocumentID = Long
+  type DocumentID = Int
   type TermWeight = Float
   type TermID = Int
